requirements-dev.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.12.13 ; python_version >= "3.12" and python_version < "4.0"
aiomysql==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.4.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
asttokens==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "4.0"
black==25.1.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.6.15 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0"
cfgv==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0"
dbf-light==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
decorator==5.2.1 ; python_version >= "3.12" and python_version < "4.0"
distlib==0.3.9 ; python_version >= "3.12" and python_version < "4.0"
executing==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.14 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.18.0 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.7.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.2.3 ; python_version >= "3.12" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.12" and python_version < "4.0"
icecream==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
identify==2.6.12 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
ipython-pygments-lexers==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
ipython==9.4.0 ; python_version >= "3.12" and python_version < "4.0"
jedi==0.19.2 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
matplotlib-inline==0.1.7 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.6.3 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
nodeenv==1.9.1 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.3.1 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
parso==0.8.4 ; python_version >= "3.12" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "emscripten"
platformdirs==4.3.8 ; python_version >= "3.12" and python_version < "4.0"
pre-commit==4.2.0 ; python_version >= "3.12" and python_version < "4.0"
prompt-toolkit==3.0.51 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.3.2 ; python_version >= "3.12" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "emscripten"
pure-eval==0.2.3 ; python_version >= "3.12" and python_version < "4.0"
pycbrf==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
pymysql==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.41 ; python_version >= "3.12" and python_version < "4.0"
stack-data==0.6.3 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.12" and python_version < "4.0"
traitlets==5.14.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.35.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
virtualenv==20.31.2 ; python_version >= "3.12" and python_version < "4.0"
watchfiles==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.12" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.20.1 ; python_version >= "3.12" and python_version < "4.0"
